update team model to have appropriate decimal places for each field

+++ server/db/models/Team.js
@@ -17,168 +17,168 @@ module.exports = db.define('team', {
     }
   },
   mp: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   fg: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   fga: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   fgp: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 3),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   thr: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   thrA: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   thrP: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 3),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   two: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   twoA: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   twoP: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 3),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ft: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ftA: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ftP: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 3),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   orb: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   drb: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   trb: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ast: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   stl: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   blk: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   tov: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   pf: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   pts: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 1),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ortg: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 2),
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   drtg: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(10, 2),
     allowNull: false,
     validate: {
       notEmpty: true
